DownloadModel to handle downloads and keep history
Currently the DownloadModel has the 'start' method to start a download
and 'cancel' to cancel it. You can see usage on DownloadManager.qml and
DownloadItem.qml. The download history is stored through QSettings and
currently I have one key to each relevant data for downloads. As there's
no support for resume yet, I don't keep progress. If you try to start
a download and the file already exists, it will be ignored. Still have
to fill the status message which is only showing download speed rate.

Reviewed-by: NOBODY