Insert basic implementation for HistoryModel.
In order to filter the results, the attribute filterString must be changed.
Then it will lookup for urls (and titles) that will have it as a substring
and present them ordered by most visited. If no url is provided, then it
will show all entries ordered by mostly recently visits.

Reviewed-by: Caio Marcelo